+++ jgiven-junit/src/test/java/com/tngtech/jgiven/junit/DataProviderTest.java
@@ -127,6 +127,25 @@ public class DataProviderTest extends ScenarioTest<GivenTestStep, WhenTestStep, 
 
     @Test
     @DataProvider( { "1", "2" } )
+    public void differences_in_nested_steps_should_be_detected( Integer methodParameter ) throws Throwable {
+        given().a_nested_step( methodParameter + 1 );
+        getScenario().finished();
+
+        ScenarioModel scenarioModel = getScenario().getModel().getLastScenarioModel();
+        if( scenarioModel.getScenarioCases().size() == 2 ) {
+            CaseArgumentAnalyser analyser = new CaseArgumentAnalyser();
+            analyser.analyze( scenarioModel );
+            ScenarioCaseModel case0 = scenarioModel.getCase( 0 );
+            assertParameter( case0, 0, "nestedStepArg" );
+
+            Word word = case0.getStep( 0 ).getNestedSteps().get( 0 ).getWords().get( 1 );
+            assertThat( word.getArgumentInfo().getParameterName() ).isEqualTo( "someIntValue" );
+        }
+
+    }
+
+    @Test
+    @DataProvider( { "1", "2" } )
     public void table_parameters_work_with_primitive_arrays( Integer arg ) {
         given().a_step_with_a_table_parameter_and_primitive_array( 1, 2, 3 );
     }